+++ NEWS.md
@@ -7,10 +7,10 @@
   * fix stratification by sex after reloading the app in some configurations
   * fix positions of signals not appearing outside other regions 
 * Shortening of long x axis labels in barcharts and adding complete axis text to tooltip
+* Reports now include a description of filter variables applied via the UI
 * Rewrite vignette to include visualization functions
 * Export plot_regional(), plot_barchart(), and plot_signals_per_week(); remove plot_agegroup_by()
 * Added option to extend the dataset up to the current day
-
 * Added new multi-seasonal harmonic method with annual and semiannual components and optional timetrend
 
 # SignalDetectionTool 0.9.0